feat(auth): integrate AppConfigService for frontend URL management and update tests accordingly refactor(auth): replace direct MailConfigService usage with AppConfigService for frontend URL retrieval
The changes introduce the AppConfigService to manage the frontend URL, enhancing the application's configuration management. This allows for a more centralized and flexible approach to configuration, making it easier to manage environment-specific settings. Additionally, the tests are updated to ensure that the new service integration is properly validated.
