+++ src/auth/auth.controller.spec.ts
@@ -1,12 +1,12 @@
 import { Test, TestingModule } from "@nestjs/testing";
 import { AuthController } from "./auth.controller";
 import { AuthService } from "./auth.service";
-
+import { AppConfigService } from "@/config/app/appConfig.service";
 import { MailConfigService } from "../config/mail/mailConfig.service";
 describe("AuthController", () => {
     let controller: AuthController;
     let authService: AuthService;
-
+    let appConfigService: AppConfigService;
     beforeEach(async () => {
         const module: TestingModule = await Test.createTestingModule({
             controllers: [AuthController],
@@ -23,11 +23,17 @@ describe("AuthController", () => {
                     provide: MailConfigService,
                     useValue: {},
                 },
+                {
+                    provide: AppConfigService,
+
+                    useValue: {},
+                },
             ],
         }).compile();
 
         controller = module.get<AuthController>(AuthController);
         authService = module.get<AuthService>(AuthService);
+        appConfigService = module.get<AppConfigService>(AppConfigService);
     });
 
     it("should be defined", () => {
@@ -35,5 +41,9 @@ describe("AuthController", () => {
     });
     it("should be defined", () => {
         expect(authService).toBeDefined();
+    });
+
+    it("should be defined", () => {
+        expect(appConfigService).toBeDefined();
     });
 });
