+++ src/auth/auth.service.spec.ts
@@ -5,7 +5,7 @@ import { GlobalService } from "../global/global.service";
 import { UsersService } from "../users/users.service";
 import { JwtService } from "@nestjs/jwt";
 import { AppConfigService } from "../config/app/appConfig.service";
-import { EmailService } from "../utils/emails/sendEmail";
+import { EmailService } from "../utils/emails/email.service";
 import { AuthConfig } from "../config/auth/auth.interface";
 
 describe("AuthService", () => {
