+++ src/config/app/app.config.ts
@@ -3,4 +3,5 @@ import { registerAs } from "@nestjs/config";
 export default registerAs("app", () => ({
     nodeEnv: process.env.NODE_ENV,
     port: process.env.PORT,
+    frontendUrl: process.env.FRONTEND_URL,
 }));
